internal/gocdk: move the "apply" command to "biome apply"; add "biome list" (#2491)
