+++ internal/cmd/gocdk/biome.go
@@ -18,6 +18,8 @@ import (
 	"bufio"
 	"context"
 	"fmt"
+	"io/ioutil"
+	"sort"
 	"strconv"
 
 	"github.com/spf13/cobra"
@@ -32,6 +34,7 @@ func registerBiomeCmd(ctx context.Context, pctx *processContext, rootCmd *cobra.
 		Short: "TODO Manage biomes",
 		Long:  "TODO more about biomes",
 	}
+
 	var launcher string
 	biomeAddCmd := &cobra.Command{
 		Use:   "add BIOME_NAME",
@@ -46,6 +49,30 @@ func registerBiomeCmd(ctx context.Context, pctx *processContext, rootCmd *cobra.
 	// prompts via flags. Maybe a JSON snippet?
 	biomeAddCmd.Flags().StringVar(&launcher, "launcher", "", "the launcher for the new biome")
 	biomeCmd.AddCommand(biomeAddCmd)
+
+	biomeListCmd := &cobra.Command{
+		Use:   "list",
+		Short: "TODO list",
+		Long:  "TODO more about listing biomes",
+		Args:  cobra.ExactArgs(0),
+		RunE: func(_ *cobra.Command, _ []string) error {
+			return biomeList(ctx, pctx)
+		},
+	}
+	biomeCmd.AddCommand(biomeListCmd)
+
+	var input bool
+	applyCmd := &cobra.Command{
+		Use:   "apply BIOME_NAME",
+		Short: "TODO Apply Terraform for BIOME",
+		Long:  "TODO more about apply",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(_ *cobra.Command, args []string) error {
+			return biomeApply(ctx, pctx, args[0], input)
+		},
+	}
+	applyCmd.Flags().BoolVar(&input, "input", true, "ask for input for Terraform variables if not directly set")
+	biomeCmd.AddCommand(applyCmd)
 
 	// TODO(rvangent): More biome subcommands: delete, list, ...?
 
@@ -64,6 +91,10 @@ type launcherData struct {
 	ServeEnabled bool
 }
 
+// biomeAdd implements the "biome add" subcommand.
+//
+// It creates a new biome, prompting the user for the launcher to use if needed,
+// as well as any launcher-specific parameters needed.
 func biomeAdd(ctx context.Context, pctx *processContext, biome, launcher string) error {
 	pctx.Logf("Adding biome %q...", biome)
 
@@ -135,5 +166,61 @@ func biomeAdd(ctx context.Context, pctx *processContext, biome, launcher string)
 		return xerrors.Errorf("gocdk biome add: %w", err)
 	}
 	pctx.Logf("Success!")
+	return nil
+}
+
+// biomeList implements the "biome list" subcommand.
+//
+// It lists the current set of biomes, computed as the subdirectories of the
+// biome rootdir.
+func biomeList(ctx context.Context, pctx *processContext) error {
+	moduleRoot, err := pctx.ModuleRoot(ctx)
+	if err != nil {
+		return xerrors.Errorf("biome list: %w", err)
+	}
+	entries, err := ioutil.ReadDir(biomesRootDir(moduleRoot))
+	if err != nil {
+		return xerrors.Errorf("biome list: %w", err)
+	}
+	var biomes []string
+	for _, entry := range entries {
+		_, err := biomeDir(moduleRoot, entry.Name())
+		if err == nil {
+			biomes = append(biomes, entry.Name())
+		}
+	}
+	sort.Strings(biomes)
+	for _, biome := range biomes {
+		pctx.Println(biome)
+	}
+	return nil
+}
+
+// biomeApply implements the "biome apply" subcommand.
+//
+// It runs "terraform init" and "terraform apply" for the named biome.
+func biomeApply(ctx context.Context, pctx *processContext, biome string, input bool) error {
+	moduleRoot, err := pctx.ModuleRoot(ctx)
+	if err != nil {
+		return xerrors.Errorf("biome apply %s: %w", biome, err)
+	}
+
+	biomePath, err := biomeDir(moduleRoot, biome)
+	if err != nil {
+		return xerrors.Errorf("biome apply %s: %w", biome, err)
+	}
+
+	c := pctx.NewCommand(ctx, biomePath, "terraform", "init", "-input="+strconv.FormatBool(input))
+	if err := c.Run(); err != nil {
+		return xerrors.Errorf("biome apply %s: %w", biome, err)
+	}
+
+	// TODO(#1821): take over steps (plan, confirm, apply) so we can
+	// dictate the messaging and errors. We should visually differentiate
+	// when we insert verbiage on top of terraform.
+	c = pctx.NewCommand(ctx, biomePath, "terraform", "apply", "-input="+strconv.FormatBool(input))
+	if err := c.Run(); err != nil {
+		return xerrors.Errorf("biome apply %s: %w", biome, err)
+	}
 	return nil
 }
