+++ internal/cmd/gocdk/testdata/recordreplay/Biome/commands.txt
@@ -4,6 +4,11 @@ gocdk biome add mybiome --> FAIL
 gocdk init myproj
 cd myproj
 
+gocdk biome list
+
 gocdk biome add mybiome --launcher local
 cd biomes
 ls mybiome
+
+gocdk biome list
+
