+++ internal/cmd/gocdk/testdata/recordreplay/Biome/out.txt
@@ -15,6 +15,9 @@ gocdk: - `gocdk biome add` to configure launch settings
 
 $ cd myproj
 
+$ gocdk biome list
+dev
+
 $ gocdk biome add mybiome --launcher local
 gocdk: Adding biome "mybiome"...
 gocdk: Success!
@@ -28,4 +31,8 @@ mybiome/
   outputs.tf
   secrets.auto.tfvars
   variables.tf
+
+$ gocdk biome list
+dev
+mybiome
