+++ internal/cmd/gocdk/testdata/recordreplay/BiomeHelp/out.txt
@@ -6,6 +6,8 @@ Usage:
 
 Available Commands:
   add         TODO Add BIOME_NAME
+  apply       TODO Apply Terraform for BIOME
+  list        TODO list
 
 Flags:
   -h, --help   help for biome
@@ -20,6 +22,8 @@ Usage:
 
 Available Commands:
   add         TODO Add BIOME_NAME
+  apply       TODO Apply Terraform for BIOME
+  list        TODO list
 
 Flags:
   -h, --help   help for biome
@@ -34,6 +38,8 @@ Usage:
 
 Available Commands:
   add         TODO Add BIOME_NAME
+  apply       TODO Apply Terraform for BIOME
+  list        TODO list
 
 Flags:
   -h, --help   help for biome
@@ -48,6 +54,8 @@ Usage:
 
 Available Commands:
   add         TODO Add BIOME_NAME
+  apply       TODO Apply Terraform for BIOME
+  list        TODO list
 
 Flags:
   -h, --help   help for biome
