+++ internal/cmd/gocdk/testdata/recordreplay/Help/out.txt
@@ -5,7 +5,6 @@ Usage:
   gocdk [command]
 
 Available Commands:
-  apply       TODO Apply Terraform for BIOME
   biome       TODO Manage biomes
   build       TODO Build a Docker image
   demo        TODO Manage demos
@@ -28,7 +27,6 @@ Usage:
   gocdk [command]
 
 Available Commands:
-  apply       TODO Apply Terraform for BIOME
   biome       TODO Manage biomes
   build       TODO Build a Docker image
   demo        TODO Manage demos
@@ -51,7 +49,6 @@ Usage:
   gocdk [command]
 
 Available Commands:
-  apply       TODO Apply Terraform for BIOME
   biome       TODO Manage biomes
   build       TODO Build a Docker image
   demo        TODO Manage demos
@@ -74,7 +71,6 @@ Usage:
   gocdk [command]
 
 Available Commands:
-  apply       TODO Apply Terraform for BIOME
   biome       TODO Manage biomes
   build       TODO Build a Docker image
   demo        TODO Manage demos
